introduce BaseObject class
