+++ src/pipecat/transports/base_transport.py
@@ -4,18 +4,16 @@
 # SPDX-License-Identifier: BSD 2-Clause License
 #
 
-import inspect
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 from typing import Optional
 
-from loguru import logger
 from pydantic import BaseModel, ConfigDict
 
 from pipecat.audio.filters.base_audio_filter import BaseAudioFilter
 from pipecat.audio.mixers.base_audio_mixer import BaseAudioMixer
 from pipecat.audio.vad.vad_analyzer import VADAnalyzer
 from pipecat.processors.frame_processor import FrameProcessor
-from pipecat.utils.utils import obj_count, obj_id
+from pipecat.utils.base_object import BaseObject
 
 
 class TransportParams(BaseModel):
@@ -43,7 +41,7 @@ class TransportParams(BaseModel):
     vad_analyzer: Optional[VADAnalyzer] = None
 
 
-class BaseTransport(ABC):
+class BaseTransport(BaseObject):
     def __init__(
         self,
         *,
@@ -51,54 +49,14 @@ class BaseTransport(ABC):
         input_name: Optional[str] = None,
         output_name: Optional[str] = None,
     ):
-        self._id: int = obj_id()
-        self._name = name or f"{self.__class__.__name__}#{obj_count(self)}"
+        super().__init__(name=name)
         self._input_name = input_name
         self._output_name = output_name
-        self._event_handlers: dict = {}
-
-    @property
-    def id(self) -> int:
-        return self._id
-
-    @property
-    def name(self) -> str:
-        return self._name
 
     @abstractmethod
     def input(self) -> FrameProcessor:
-        raise NotImplementedError
+        pass
 
     @abstractmethod
     def output(self) -> FrameProcessor:
-        raise NotImplementedError
-
-    def event_handler(self, event_name: str):
-        def decorator(handler):
-            self.add_event_handler(event_name, handler)
-            return handler
-
-        return decorator
-
-    def add_event_handler(self, event_name: str, handler):
-        if event_name not in self._event_handlers:
-            raise Exception(f"Event handler {event_name} not registered")
-        self._event_handlers[event_name].append(handler)
-
-    def _register_event_handler(self, event_name: str):
-        if event_name in self._event_handlers:
-            raise Exception(f"Event handler {event_name} already registered")
-        self._event_handlers[event_name] = []
-
-    async def _call_event_handler(self, event_name: str, *args, **kwargs):
-        try:
-            for handler in self._event_handlers[event_name]:
-                if inspect.iscoroutinefunction(handler):
-                    await handler(self, *args, **kwargs)
-                else:
-                    handler(self, *args, **kwargs)
-        except Exception as e:
-            logger.exception(f"Exception in event handler {event_name}: {e}")
-
-    def __str__(self):
-        return self.name
+        pass
